backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
supabase==1.2.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.13
staticmap==0.4.0
Pillow==10.0.1
polyline==2.0.0
